Move resource limits into separate file
